requirements.txt
fastapi
uvicorn
cryptography
pillow
numpy
python-multipart

